move form logic to formutils file
