+++ src/utils/formUtils.ts
@@ -6,7 +6,7 @@ import {
   TransducerType,
 } from '../data/data';
 
-export interface FormState {
+interface FormState {
   name: string;
   location: string;
   department: string;
@@ -21,8 +21,83 @@ export interface FormState {
   service: boolean;
 }
 
+export const initialState: FormState = {
+  name: '',
+  location: '',
+  department: '',
+  room: '',
+  type: '',
+  serial: '',
+  internal: '',
+  control: '',
+  received: '',
+  condition: '',
+  notes: '',
+  service: false
+};
+
+
+type Type = 'CHANGE_INPUT' | 'CHANGE_CHECKBOX' | 'RESET';
+
+type Action = {
+  type: Type,
+  payload: {
+    name?: string;
+    value?: string;
+    checked?: boolean;
+    initialState?: FormState
+  };
+};
+
+export const reducer = (state: FormState, action: Action): FormState => {
+  switch(action.type) {
+    case 'CHANGE_INPUT':
+      if (action.payload.name === 'condition') {
+        if (action.payload.value === 'Broken (Out of Service)') {
+          return {
+            ...state,
+            [action.payload.name]: action.payload.value,
+            ['service']: true
+          };
+        } else {
+          return {
+            ...state,
+            [action.payload.name]: action.payload.value,
+            ['service']: false
+          };
+        }
+      }
+
+      return {
+        ...state,
+        [action.payload.name]: action.payload.value
+      };
+
+    case 'CHANGE_CHECKBOX':
+      if (action.payload.checked) {
+        return {
+          ...state,
+          [action.payload.name]: action.payload.checked,
+          ['condition']: 'Broken (Out of Service)'
+        };
+      }
+
+      return {
+        ...state,
+        [action.payload.name]: action.payload.checked,
+        ['condition']: 'Working'
+      };
+
+    case 'RESET':
+      return initialState;
+
+    default: 
+      return state;
+  };
+};
+
 // Create a new transducer object from inputted form data
-export const createTransducerObject = (formData: FormState): Transducer => {
+export const createTransducer = (formData: FormState): Transducer => {
   const newTransducer: Transducer = {
     id: crypto.randomUUID(),
     name: formData.name,
